Feat: Created object generator module and relocated methods from stat_tracker
