+++ lib/stat_tracker2.rb
@@ -3,46 +3,21 @@ require './lib/game'
 require './lib/game_team'
 require './lib/team'
 require './lib/league'
+require './lib/object_generator'
 
 class StatTracker
-  attr_reader :games,
-              :teams,
-              :game_teams,
-              :league
+  include ObjectGenerator
+  attr_reader :league
 
   def self.from_csv(locations)
     new(locations)
   end
 
   def initialize(locations)
-    @games = games_array_creator(locations[:games])
-    @teams = teams_array_creator(locations[:teams])
-    @game_teams = game_teams_array_creator(locations[:game_teams])
-    @league = League.new(@games, @teams, @game_teams)
-  end
-
-  def games_array_creator(path)
-    games_array = []
-    CSV.foreach(path, headers: true, header_converters: :symbol) do |row|
-      games_array << Game.new(row)
-    end
-    games_array
-  end
-
-  def teams_array_creator(path)
-    teams_array = []
-    CSV.foreach(path, headers: true, header_converters: :symbol) do |row|
-      teams_array << Team.new(row)
-    end
-    teams_array
-  end
-
-  def game_teams_array_creator(path)
-    game_teams_array = []
-    CSV.foreach(path, headers: true, header_converters: :symbol) do |row|
-      game_teams_array << GameTeam.new(row)
-    end
-    game_teams_array
+    games = games_array_creator(locations[:games])
+    teams = teams_array_creator(locations[:teams])
+    game_teams = game_teams_array_creator(locations[:game_teams])
+    @league = League.new(games, teams, game_teams)
   end
 
 #Game Statistics Methods
@@ -55,27 +30,15 @@ class StatTracker
   end
 
   def percentage_home_wins
-
-    (@league.home_wins_counter.to_f / @games.count).round(2)
-
-
-
+    (@league.home_wins_counter.to_f / @league.all_games.count).round(2)
   end
 
   def percentage_visitor_wins
-
-    (@league.visitor_wins_counter.to_f / @games.count).round(2)
-
-
-
+    (@league.visitor_wins_counter.to_f / @league.all_games.count).round(2)
   end
 
   def percentage_ties
-
-    (@league.tie_counter.to_f / @games.count).round(2)
-
-
-
+    (@league.tie_counter.to_f / @league.all_games.count).round(2)
   end
 
   def count_of_games_by_season
@@ -87,11 +50,7 @@ class StatTracker
   end
 
   def average_goals_by_season
-
     @league.avg_goals_by_season
-
-
-
   end
   #League Statistics Methods
   def count_of_teams
@@ -109,51 +68,35 @@ class StatTracker
   end
 
   def highest_scoring_visitor
-
     max = @league.avg_away_team_by_goals.max_by{|team_id, avg_goals| avg_goals}
     @league.team_id_to_team_name(max[0])
-
-
   end
 
   def lowest_scoring_visitor
-
     min = @league.avg_away_team_by_goals.min_by{|team_id, avg_goals| avg_goals}
     @league.team_id_to_team_name(min[0])
-
-
   end
 
   def highest_scoring_home_team
-
     max = @league.avg_home_team_by_goals.max_by{|team_id, avg_goals| avg_goals}
     @league.team_id_to_team_name(max[0])
-
-
   end
 
   def lowest_scoring_home_team
-
-
     min = @league.avg_home_team_by_goals.min_by{|team_id, avg_goals| avg_goals}
     @league.team_id_to_team_name(min[0])
-
   end
 #Season Statistics Methods
   def winningest_coach(season)
-
     @league.coaches_by_win_percentage(season).max_by do |coach, win_percentage|
       win_percentage
     end[0]
-
   end
 
   def worst_coach(season)
-
     @league.coaches_by_win_percentage(season).min_by do |coach, win_percentage|
       win_percentage
     end[0]
-
   end
 
   def most_accurate_team(season)
@@ -161,7 +104,6 @@ class StatTracker
       accuracy
     end[0]
     @league.team_id_to_team_name(team_id)
-
   end
 
   def least_accurate_team(season)
@@ -169,7 +111,6 @@ class StatTracker
       accuracy
     end[0]
     @league.team_id_to_team_name(team_id)
-
   end
 
   def most_tackles(season)
@@ -177,7 +118,6 @@ class StatTracker
       tackles
     end[0]
     @league.team_id_to_team_name(team_id)
-
   end
 
   def fewest_tackles(season)
@@ -185,54 +125,36 @@ class StatTracker
       tackles
     end[0]
     @league.team_id_to_team_name(team_id)
-
   end
 #Team Statistics Methods
   def team_info(given_team_id)
     selected_team = @league.find_team(given_team_id)
     @league.display_team_info(selected_team)
-
-
-
   end
 
   def best_season(given_team_id)
     @league.seasons_by_wins(given_team_id).max_by do |season, win_percentage|
       win_percentage
     end[0]
-
-
   end
 
   def worst_season(given_team_id)
     @league.seasons_by_wins(given_team_id).min_by do |season, win_percentage|
       win_percentage
     end[0]
-
-
   end
 
   def average_win_percentage(given_team_id)
     team_tally = @league.wins_losses_tally(given_team_id)
     (team_tally[:wins].to_f / team_tally[:total_games]).round(2)
-
-
-
   end
 
   def most_goals_scored(team_id)
     @league.goals_scored_in_game(team_id).max
-
-
-
-
   end
 
   def fewest_goals_scored(team_id)
     @league.goals_scored_in_game(team_id).min
-
-
-
   end
 
   def favorite_opponent(given_team_id)
@@ -240,7 +162,6 @@ class StatTracker
       win_percentage
     end[0]
     @league.team_id_to_team_name(fav_opponent_team_id)
-
   end
 
   def rival(given_team_id)
@@ -248,7 +169,5 @@ class StatTracker
       win_percentage
     end[0]
     @league.team_id_to_team_name(rival_team_id)
-
   end
-
 end
